requirements.txt
appnope==0.1.0
backcall==0.1.0
decorator==4.3.0
Django==2.0.7
django-environ==0.4.5
django-filter==2.0.0
djangorestframework==3.8.2
gunicorn==19.9.0
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
